Sender pdf og vedlegg til mottak slik at de kan journalføres derfra. Fikser feil som gjorde at appen ikke klarte å kjøre opp lokalt
